Split `io.py` into multiple files
